# Conflicts:
#	.gitignore
#	Cargo.lock
#	examples/Cargo.lock
#	libwebrtc/src/encoded_audio_frame.rs
#	libwebrtc/src/encoded_audio_frame_stream.rs
#	libwebrtc/src/encoded_video_frame.rs
#	libwebrtc/src/encoded_video_frame_stream.rs
#	libwebrtc/src/native/encoded_audio_frame_stream.rs
#	libwebrtc/src/native/encoded_video_frame_stream.rs
#	libwebrtc/src/native/mod.rs
#	libwebrtc/src/native/sender_report_stream.rs
#	libwebrtc/src/sender_report.rs
#	libwebrtc/src/sender_report_stream.rs
#	livekit-protocol/protocol
#	livekit/src/room/mod.rs
#	livekit/src/room/participant/remote_participant.rs
#	livekit/src/room/track/remote_video_track.rs
#	webrtc-sys/build.rs
#	webrtc-sys/libwebrtc/build_macos.sh
#	webrtc-sys/src/lib.rs

+++ examples/save_to_disk/src/main.rs
@@ -95,6 +95,7 @@ async fn main() {
     println!("Connected to room: {} - {}", room.name(), room.sid());
 
     while let Some(msg) = rx.recv().await {
+        #[allow(clippy::single_match)]
         match msg {
             RoomEvent::TrackSubscribed {
                 track,
@@ -127,7 +128,7 @@ async fn record_track(audio_track: RemoteAudioTrack) -> Result<(), std::io::Erro
     let mut wav_writer = WavWriter::create(FILE_PATH, header).await?;
     let mut audio_stream = NativeAudioStream::new(rtc_track);
 
-    let max_record = 5 * header.sample_rate * header.num_channels as u32;
+    let max_record = 5 * header.sample_rate * header.num_channels;
     let mut sample_count = 0;
     'recv_loop: while let Some(frame) = audio_stream.next().await {
         let data = resampler.remix_and_resample(
